// frontend/src/types/Entertainer.ts
export interface Entertainer {
    entertainerID: number;
    entStageName: string;
    entSSN: string;
    entStreetAddress: string;
    entCity: string;
    entState: string;
    entZipCode: number;
    entPhoneNumber: string;
    entWebPage: string;
    entEMailAddress: string;
    dateEntered: string;
    bookingCount: number;
    mostRecentBooking: string
}